+++ src/backoff_strategies.rs
@@ -19,7 +19,7 @@ pub trait BackoffStrategy<E> {
 
 /// No backoff. This will make the future be retried immediately without any delay in between
 /// attempts.
-#[derive(Debug)]
+#[derive(Debug, Clone, Copy)]
 pub struct NoBackoff;
 
 impl<E> BackoffStrategy<E> for NoBackoff {
@@ -32,7 +32,7 @@ impl<E> BackoffStrategy<E> for NoBackoff {
 }
 
 /// Exponential backoff. The delay will double each time.
-#[derive(Debug)]
+#[derive(Debug, Clone, Copy)]
 pub struct ExponentialBackoff {
     pub(crate) delay: Duration,
 }
@@ -49,7 +49,7 @@ impl<E> BackoffStrategy<E> for ExponentialBackoff {
 }
 
 /// Fixed backoff. The delay wont change between attempts.
-#[derive(Debug)]
+#[derive(Debug, Clone, Copy)]
 pub struct FixedBackoff {
     pub(crate) delay: Duration,
 }
@@ -64,7 +64,7 @@ impl<E> BackoffStrategy<E> for FixedBackoff {
 }
 
 /// Linear backoff. The delay will scale linearly with the number of attempts.
-#[derive(Debug)]
+#[derive(Debug, Clone, Copy)]
 pub struct LinearBackoff {
     pub(crate) delay: Duration,
 }
@@ -79,6 +79,7 @@ impl<E> BackoffStrategy<E> for LinearBackoff {
 }
 
 /// A custom backoff strategy defined by a function.
+#[derive(Clone, Copy)]
 pub struct CustomBackoffStrategy<F> {
     pub(crate) f: F,
 }